toc2: sidebar + html export
