+++ nbextensions/usability/toc2/toc2.yaml
@@ -6,23 +6,27 @@ Icon: icon.png
 Main: main.js
 Compatibility: 4.x
 Parameters:
-- name: toc_number_sections
+- name: number_sections
   description: Automatically number notebook's sections
   input_type: checkbox
   default: true
-- name: toc_threshold
+- name: threshold
   description: Maximum level of nested sections to display on the tables of contents
   input_type: number
   min: -1
   step: 1
   default: 4
 - name: toc_cell
-  description: Add a Table of Contents at the top of the notebook
+  description: Add a Table of Contents cell at the top of the notebook
   input_type: checkbox
   default: false
 - name: toc_window_display
   description: Display toc window at startup
   input_type: checkbox
   default: false
+- name: sideBar
+  description: Display Table of Contents as a sidebar (otherwise as a floating window)
+  input_type: checkbox
+  default: false
